refactor: 공통 컴포넌트명 변경 (Text.tsx -> Typography.tsx)
